+++ tests/integration/test_spans.py
@@ -3,7 +3,10 @@ import json
 import uuid
 from collections import Counter
 from datetime import datetime, timedelta, timezone, UTC
-from .consts import TRANSACTION_EXTRACT_MIN_SUPPORTED_VERSION
+from .consts import (
+    TRANSACTION_EXTRACT_MIN_SUPPORTED_VERSION,
+    METRICS_EXTRACTION_MIN_SUPPORTED_VERSION,
+)
 
 import pytest
 from opentelemetry.proto.common.v1.common_pb2 import AnyValue, KeyValue
@@ -2045,3 +2048,125 @@ def test_dynamic_sampling(
 
     spans_consumer.assert_empty()
     outcomes_consumer.assert_empty()
+
+
+def test_metrics_summary_with_extracted_spans(
+    mini_sentry,
+    relay_with_processing,
+    metrics_summaries_consumer,
+):
+    mini_sentry.global_config["options"] = {
+        "relay.compute-metrics-summaries.sample-rate": 1.0
+    }
+
+    metrics_summaries_consumer = metrics_summaries_consumer()
+
+    relay = relay_with_processing()
+    project_id = 42
+    project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"]["features"] = [
+        "organizations:custom-metrics",
+        "projects:span-metrics-extraction",
+        "organizations:indexed-spans-extraction",
+    ]
+    project_config["config"]["transactionMetrics"] = {
+        "version": TRANSACTION_EXTRACT_MIN_SUPPORTED_VERSION,
+    }
+    project_config["config"]["metricExtraction"] = {
+        "version": METRICS_EXTRACTION_MIN_SUPPORTED_VERSION,
+        "metrics": [
+            {
+                "category": "span",
+                "mri": "d:custom/my_metric@millisecond",
+                "field": "span.duration",
+            }
+        ],
+    }
+
+    event = make_transaction({"event_id": "cbf6960622e14a45abc1f03b2055b186"})
+    end = datetime.now(timezone.utc) - timedelta(seconds=1)
+    duration = timedelta(milliseconds=500)
+    start = end - duration
+    event["spans"] = [
+        {
+            "description": "GET /api/0/organizations/?member=1",
+            "op": "http",
+            "origin": "manual",
+            "parent_span_id": "aaaaaaaaaaaaaaaa",
+            "span_id": "bbbbbbbbbbbbbbbb",
+            "start_timestamp": start.isoformat(),
+            "status": "success",
+            "timestamp": end.isoformat(),
+            "trace_id": "ff62a8b040f340bda5d830223def1d81",
+        },
+    ]
+
+    mri = "c:spans/some_metric@none"
+    metrics_summary = {
+        mri: [
+            {
+                "min": 1.0,
+                "max": 2.0,
+                "sum": 3.0,
+                "count": 4,
+                "tags": {
+                    "environment": "test",
+                },
+            },
+        ],
+    }
+    event["_metrics_summary"] = metrics_summary
+
+    relay.send_event(project_id, event)
+
+    metrics_summaries = metrics_summaries_consumer.get_metrics_summaries(
+        timeout=10.0, n=3
+    )
+    expected_mris = ["c:spans/some_metric@none", "d:custom/my_metric@millisecond"]
+    for metric_summary in metrics_summaries:
+        assert metric_summary["mri"] in expected_mris
+
+
+def test_metrics_summary_with_standalone_spans(
+    mini_sentry,
+    relay_with_processing,
+    metrics_summaries_consumer,
+):
+    mini_sentry.global_config["options"] = {
+        "relay.compute-metrics-summaries.sample-rate": 1.0
+    }
+
+    metrics_summaries_consumer = metrics_summaries_consumer()
+
+    relay = relay_with_processing()
+    project_id = 42
+    project_config = mini_sentry.add_full_project_config(project_id)
+    project_config["config"]["features"] = [
+        "projects:span-metrics-extraction",
+        "organizations:standalone-span-ingestion",
+    ]
+    project_config["config"]["metricExtraction"] = {
+        "version": METRICS_EXTRACTION_MIN_SUPPORTED_VERSION,
+        "metrics": [
+            {
+                "category": "span",
+                "mri": "d:custom/my_metric@millisecond",
+                "field": "span.duration",
+            }
+        ],
+    }
+
+    duration = timedelta(milliseconds=500)
+    now = datetime.now(timezone.utc)
+    end = now - timedelta(seconds=1)
+    start = end - duration
+
+    envelope = envelope_with_spans(start, end)
+    relay.send_envelope(project_id, envelope)
+
+    metrics_summaries = metrics_summaries_consumer.get_metrics_summaries(
+        timeout=10.0, n=4
+    )
+    expected_mris = ["c:spans/some_metric@none", "d:custom/my_metric@millisecond"]
+    for metric_summary in metrics_summaries:
+        assert metric_summary["mri"] in expected_mris