+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Ingest profiler_id in the profile context and in spans. ([#3714](https://github.com/getsentry/relay/pull/3714), [#3784](https://github.com/getsentry/relay/pull/3784))
 - Support extrapolation of metrics extracted from sampled data, as long as the sample rate is set in the DynamicSamplingContext. ([#3753](https://github.com/getsentry/relay/pull/3753))
 - Extract thread ID and name in spans. ([#3771](https://github.com/getsentry/relay/pull/3771))
+- Compute metrics summary on the extracted custom metrics. ([#3769](https://github.com/getsentry/relay/pull/3769))
+
+>>>>>>> master
 
 ## 24.6.0